requirements.txt
supabase>=2.10.0
customtkinter>=5.2.0
pandas>=2.1.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
pydantic>=2.4.0
pillow>=10.0.0
requests>=2.31.0
